gcloud-k8s-backup/requirements.txt
python-dateutil==2.6.1
pytz==2018.3
google-api-python-client==1.6.5
google-auth-httplib2==0.0.3
